# api/docs/swagger.yaml
definitions:
...
  model.CreateStudentReq:
    properties:
      lastname:
        type: string
      name:
        type: string
      phone_number:
        type: string
      subject1:
        type: string
      subject2:
        type: string
    type: object
  model.CreateStudentResp:
    properties:
      student_id:
        type: string
    type: object
  model.CreateSubjectReq:
    properties:
      name:
        type: string
    type: object
...
  model.Error:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  model.GetStudentResultResp:
    properties:
      results:
        items:
          $ref: '#/definitions/model.StudentResult'
        type: array
    type: object
  model.GetStudentsResp:
    properties:
      students:
        items:
          $ref: '#/definitions/model.Student'
        type: array
    type: object
  model.GetStudentsResultResp:
    properties:
      count:
        type: integer
      students_results:
        items:
          $ref: '#/definitions/model.StudentReslt'
        type: array
    type: object
  model.GetSubjectsResp:
    properties:
      subjects:
        items:
          $ref: '#/definitions/model.Subject'
        type: array
    type: object
...
  model.QuestionResult:
    properties:
      number:
        type: integer
      status:
        type: boolean
    type: object
...
  model.Status:
    properties:
      message:
        type: string
    type: object
  model.Student:
    properties:
      id:
        type: string
      lastname:
        type: string
      name:
        type: string
      phone_number:
        type: string
      student_id:
        type: string
      subject1:
        type: string
      subject2:
        type: string
    type: object
  model.StudentReslt:
    properties:
      ball:
        type: number
      day:
        type: string
      lastname:
        type: string
      name:
        type: string
      result:
        items:
          $ref: '#/definitions/model.QuestionResult'
        type: array
      student_id:
        type: string
      subject1:
        type: string
      subject2:
        type: string
    type: object
  model.StudentResult:
    properties:
      ball:
        type: number
      result:
        items:
          $ref: '#/definitions/model.QuestionResult'
        type: array
      template_id:
        type: string
    type: object
  model.StudentsStatus:
    properties:
      correct:
        type: integer
      incorrect:
        type: integer
      incorrect_students:
        items:
          $ref: '#/definitions/model.Student'
        type: array
    type: object
  model.Subject:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  model.UpdateStudentReq:
    properties:
      lastname:
        type: string
      name:
        type: string
      phone_number:
        type: string
      subject1:
        type: string
      subject2:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /students:
    get:
      description: Berilgan ID bo‘yicha talabani yoki barcha talabalarni olish
      parameters:
      - description: Talabaning unikal ID'si (Majburiy emas, agar berilmasa barcha
          talabalar qaytariladi)
        in: query
        name: id
        type: string
      responses:
        "200":
          description: Talabalar ma'lumotlari
          schema:
            $ref: '#/definitions/model.GetStudentsResp'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Talabalar ma'lumotlarini olish
      tags:
      - Students
  /students/{student_id}/result:
    get:
      consumes:
      - application/json
      description: Retrieves the result of a student based on student_id and template_id
      parameters:
      - description: Student ID
        in: path
        name: student_id
        required: true
        type: string
      - description: Template ID
        in: query
        name: template_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetStudentResultResp'
        "500":
          description: Serverda xatolik
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Get student result
      tags:
      - Students
  /students/create:
    post:
      consumes:
      - application/json
      description: Yangi talabani ro‘yxatga olish
      parameters:
      - description: Talaba ma'lumotlari
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateStudentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateStudentResp'
        "400":
          description: Noto'g'ri ma'lumot kiritildi
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Talaba yaratish
      tags:
      - Students
  /students/delete/{id}:
    delete:
      description: Berilgan ID bo‘yicha talabani o‘chirish
      parameters:
      - description: Talabaning unikal ID'si
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Muvaffaqiyatli o‘chirildi
          schema:
            $ref: '#/definitions/model.Status'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Talabani o'chirish
      tags:
      - Students
  /students/results:
    get:
      consumes:
      - application/json
      description: Berilgan kun va fanlarga mos keluvchi talabalarning natijalarini
        qaytaradi
      parameters:
      - description: Natija kuni
        in: query
        name: day
        type: string
      - description: Birinchi fan ID si
        in: query
        name: subject1_id
        type: string
      - description: Ikkinchi fan ID si
        in: query
        name: subject2_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Talabalarning natijalari
          schema:
            $ref: '#/definitions/model.GetStudentsResultResp'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Talabalarning natijalarini olish
      tags:
      - Students
  /students/update/{id}:
    put:
      consumes:
      - application/json
      description: Berilgan ID bo‘yicha talabani yangilash
      parameters:
      - description: Talabaning unikal ID'si
        in: path
        name: id
        required: true
        type: string
      - description: Yangilangan talaba ma'lumotlari
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateStudentReq'
      produces:
      - application/json
      responses:
        "200":
          description: Muvaffaqiyatli yangilandi
          schema:
            $ref: '#/definitions/model.Status'
        "400":
          description: Noto'g'ri ma'lumot kiritildi
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Talaba ma'lumotlarini yangilash
      tags:
      - Students
  /students/upload:
    post:
      consumes:
      - multipart/form-data
      description: Excel fayl orqali talabalar ma'lumotlarini yuklash
      parameters:
      - description: Excel fayl (XLSX format)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Talabalar muvaffaqiyatli yuklandi
          schema:
            $ref: '#/definitions/model.StudentsStatus'
        "400":
          description: Fayl yuklashda xatolik
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Server xatosi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Talabalar ro'yxatini yuklash
      tags:
      - Files
  /subjects/create:
    post:
      consumes:
      - application/json
      description: Yangi fan qo‘shish
      parameters:
      - description: Yangi fan ma'lumotlari
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: Muvaffaqiyatli qo‘shildi
          schema:
            $ref: '#/definitions/model.Status'
        "400":
          description: Noto'g'ri ma'lumot kiritildi
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Fan yaratish
      tags:
      - Subjects
  /subjects/get:
    get:
      description: Agar ID berilsa, shu fanni, aks holda barcha fanlarni qaytaradi
      parameters:
      - description: Fan ID'si (Majburiy emas, agar berilmasa barcha fanlar qaytariladi)
        in: query
        name: id
        type: string
      responses:
        "200":
          description: Fanlar ma'lumotlari
          schema:
            $ref: '#/definitions/model.GetSubjectsResp'
        "500":
          description: Serverda xatolik yuz berdi
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Fan(lar) ma'lumotlarini olish
      tags:
      - Subjects
  /subjects/update/{id}:
    put:
      consumes:
      - application/json
      description: Berilgan ID bo‘yicha fanni yangilaydi
      parameters:
      - description: Fan IDsi
        in: path
        name: id
        required: true
        type: string
      - description: Yangi fan nomi
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Status'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      security:
      - BearerAuth: []
      summary: Fan nomini yangilash
      tags:
      - Subjects
...
securityDefinitions:
  BearerAuth:
    description: 'Tokenni faqat o‘zi yozing, masalan: eyJhbGciOiJIUzI1NiIs..'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"